Entropy and New Algorithms
I realised that changing the function set_applicable_actions(actions, predicates) was NOT a good idea. Cleaning the effects of the applicable actions, in accordance with the predicates, means more memory to store different action objects. Besides, the algorithm wastes a lot of time doing than for some states that are already in the graph so all this effort is going to be discarded by the function check_state(s). However, the idea was useful because this task is necessary to compute the entropy of a state.

This commit includes: first the methods that compute the entropy of a state object and finally a set of solvers that, in theory, are able to solve any of the validated problems. Let's see.